refactor the file structure, added components directory to contain components
