+++ frontend/pages/_app.tsx
@@ -1,6 +1,6 @@
+import Navigation from "@/components/homepage/layout/Navigation";
 import "@/styles/globals.css";
 import type { AppProps } from "next/app";
-import Navigation from "./components/Navigation/Navigation";
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
