+++ frontend/pages/index.tsx
@@ -1,6 +1,4 @@
-import Image from "next/image";
 import { Inter } from "next/font/google";
-import Navigation from "./components/Navigation/Navigation";
 
 const inter = Inter({ subsets: ["latin"] });
